Does having a covered stadium makes an advantage
I wanted to check, if it would make sense for a NFL team, to build a stadium that is covering the field to have an advantage at home.
There is no significant outcome, but it tends, to look, that NFL teams without a covered roof, tend to play better in closed stadiums, than teams with a closed stadium, when they play in open stadiums.